+++ examples/service.js
@@ -4,13 +4,13 @@ var geo = require('../lib/geosearch');
 var ip = process.argv[2] || '8.8.8.8';
 var options = {
     services: {
-        'maxmind-dat': false,
-        'maxmind-mmdb': 1,
+        'maxmind-dat': 1,
+        'maxmind-mmdb': false,
         'ipgeobase': false,
         'geobytes': false,
         'freegeoip': false,
-        'ip-api':false,
-        'ipinfo':false
+        'ip-api': false,
+        'ipinfo': false
     }
 };
 geo.setOptions(options);