+++ test/services/maxmind-dat.js
@@ -3,6 +3,7 @@
 const path = require('path');
 const expect = require("chai").expect;
 const service = require(path.join(__dirname, "../../lib/services/maxmind-dat"));
+//service.setParam('directory','/var/www1/');
 
 describe('Testing maxmind-dat', function () {
     before(function () {
@@ -10,7 +11,6 @@ describe('Testing maxmind-dat', function () {
     });
     it("basic ip check, result should be object, err = null", function (done) {
         service.lookup('8.8.8.8', function (err, result) {
-            console.log(result);
             try {
                 expect(result).be.a('object');
                 done();